+++ board-frontend/src/scripts/FileUpload.js
@@ -1,3 +1,7 @@
+import {activeDocument, getDt, updateCanvas} from "./Document.js";
+import {CanvasObjectType} from "./canvas/CanvasObjectType.js";
+import CanvasObjectWrapper from "./canvas/CanvasObjectWrapper.js";
+import CanvasImage from "./canvas/canvasObjects/CanvasImage.js";
 const fileUploadButton = document.getElementById("fileUploadButton");
 
 fileUploadButton.addEventListener('change', (event) => {
@@ -16,4 +20,9 @@ function newFile(file) {
         };
     })(image);
     reader.readAsDataURL(file);
+
+    let object = CanvasImage.create(0, 0, image);
+    let id = (Math.random() - 0.5) * 32000;
+    updateCanvas.update(CanvasObjectType.IMAGE, id, CanvasObjectWrapper.create(getDt(), object));
+    activeDocument.canvas.insert(CanvasObjectType.IMAGE, id, object);
 }